Discord_bot/requirements.txt
discord.py==2.2.2
openai==0.27.4
python-dotenv==1.0.0
requests==2.28.2
pytest==7.3.0
black==23.3.0
flake8==6.0.0
isort==5.12.0
pytest-cov==4.0.0
pre-commit==3.2.2
loguru==0.7.0
pylint==2.17.3
pytest-mock==3.10.0
pytest-asyncio==0.21.0
icalendar==5.0.5
